+++ main/F446-INUWASHI/src/device/actuator/sts3032.h
@@ -15,6 +15,8 @@ class STS3032 {
 
     const int maximumSpeed = 7000;
     const int baudRate = 1000000;
+
+    const int DefaultSpeed = 100;
     
     void directDrive(int id, int percent, int acceleration = 0);
     void drive(int velocity, int angle);
