+++ main/F446-INUWASHI/src/neko/neko.cpp
@@ -1,3 +1,6 @@
 #include "./neko.h"
 
 Location location;
+Field Map[200];
+
+VICTIM victim;
