--- main/F446-INUWASHI/src/rtos/algorithm/drive.h
@@ -1,9 +1,0 @@
-#ifndef _DRIVE_H
-#define _DRIVE_H
-#include "Arduino.h"
-class DRIVE {
-   public:
-   private:
-};
-
-#endif
